SketchMLP/requirements.txt
cupy_cuda12x
einops==0.4.1
h5py==3.8.0
lmdb==1.4.0
numpy==1.21.5
opencv_python==4.6.0.66
Pillow==9.2.0
rdp==0.8
scikit_image==0.19.3
scipy==1.7.3
tensorboardX==2.2
timm==0.6.12
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
